Resetting pin after momentary HIGH outside the ISR
When the chip is instructed to momentarily pull a pin high, a counter
is set to a configurable value. This counter is then decreased in the
ISR. Up until now, the pins were pulled low inside the ISR. This is
bad practice as we should aim to keep the ISR as short as possible.

This commit addresses this issue by merely setting a flag in the ISR;
the pins are pulled low in the regular event loop.